please update all the readme files with all the new features, and update the technical document thorughly too. also i need to implement SSH and other features into this workflow. basically what happen is this,
1. I import all my dictionaries
2. in the server config i added my radius server and put all the commands to check radius config and avilablitiy of radius tools i need. when I click on "Test connection" system actually needs to SSH into the given user correctly and run all the commands and if it got expected results for all status will be connected
3. do the same for SQL server
2. I create few packets from packet editor
3. using scenario builder im going to build my whole scenario.
4. then im going to run the scenario.
when i run the scenario this system acutally need to ssh into the configed server in scenario builder, send given packets using radtest or radclient with all the custom attributes i enter, see if we get expected result, if we got expected result move into next step.

lets say next step is a sql query. in the scenario builder for that SQL query system need to loginto or run the query to that given SQL server somehow and check the reuslt is same as expected result if yes move to next one. like that we need to test all the things.

rightnow this system do not have SSH or SQL server loging permission. so i need complete a to z documentaion about how to develeop this system backend to it. what codes and code lined to edit, what filed to edit what to do how to do everything. so i can locally develop the system and complete the whole process.